database: make the InitTestDB function available to use from other packages
